refactor: SocialAccountWriter 파일 병합 (#1213)
